Ajouts des methodes manquantes du crud dans tous les micros

+++ parkings/src/main/java/fr/uphf/parkings/entities/Parking.java
@@ -17,7 +17,7 @@ import lombok.*;
 public class Parking {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idParking;
+    private Long idParking;
 
     @NotBlank
     private String nomZoneParking;
